Rework ResponseReferencesPlugin
