+++ flask_smorest/spec/constants.py
@@ -1,0 +1,4 @@
+"""Constants"""
+
+DEFAULT_REQUEST_BODY_CONTENT_TYPE = 'application/json'
+DEFAULT_RESPONSE_CONTENT_TYPE = 'application/json'
